+++ src/rdkafka_queue.c
@@ -742,10 +742,25 @@ rd_kafka_queue_t *rd_kafka_queue_get_partition (rd_kafka_t *rk,
 }
 
 rd_kafka_queue_t *rd_kafka_queue_get_background (rd_kafka_t *rk) {
-        if (rk->rk_background.q)
-                return rd_kafka_queue_new0(rk, rk->rk_background.q);
-        else
-                return NULL;
+        rd_kafka_queue_t *rkqu;
+
+        rd_kafka_wrlock(rk);
+        if (!rk->rk_background.q) {
+                char errstr[256];
+
+                if (rd_kafka_background_thread_create(rk,
+                                                      errstr, sizeof(errstr))) {
+                        rd_kafka_log(rk, LOG_ERR, "BACKGROUND",
+                                     "Failed to create background thread: %s",
+                                     errstr);
+                        rd_kafka_wrunlock(rk);
+                        return NULL;
+                }
+        }
+
+        rkqu = rd_kafka_queue_new0(rk, rk->rk_background.q);
+        rd_kafka_wrunlock(rk);
+        return rkqu;
 }